backend/pyproject.toml
[tool.poetry]
name = "sse"
version = "0.1.0"
description = ""
authors = ["hosimesi <hosimesi11@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"


[tool.poetry.group.chat.dependencies]
langchain = "^0.0.327"
openai = "^1.1.1"
gunicorn = "^21.2.0"
fastapi = "^0.104.1"
grpcio = "^1.59.2"
crashtest = "^0.4.1"
sse-starlette = "^1.6.5"
coloredlogs = "^15.0.1"
uvicorn = "^0.15.0"
grpcio-tools = "^1.59.3"
taskipy = "^1.12.0"
pydantic = "^2.5.1"
pydantic-settings = "^2.1.0"
mypy-protobuf = "^3.5.0"
websockets = "^12.0"


[tool.taskipy.tasks]
update = { cmd = "poetry export --without-hashes --with chat --output requirements/chat.txt", help = "update requirements" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
